+++ src/modules/plugins/plugins.service.ts
@@ -79,6 +79,11 @@ export class PluginsService {
 
   // verified plugins cache
   private verifiedPlugins: string[] = [];
+  private verifiedPluginsIcons: { [key: string]: string } = {};
+  private verifiedPluginsIconsPrefix = 'https://raw.githubusercontent.com/homebridge/verified/latest/';
+
+  private verifiedPluginsJson = 'https://raw.githubusercontent.com/homebridge/verified/latest/verified-plugins.json';
+  private verifiedPluginsIconsJson = 'https://raw.githubusercontent.com/homebridge/verified/latest/plugin-icons.json';
 
   // misc schemas
   private miscSchemas = {
@@ -289,11 +294,13 @@ export class PluginsService {
         plugin.latestVersion = pkg.package.version;
         plugin.lastUpdated = pkg.package.date;
         plugin.description = (pkg.package.description) ?
-          pkg.package.description.replace(/(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pkg.package.name;
+          pkg.package.description.replace(/\(?(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pkg.package.name;
         plugin.links = pkg.package.links;
         plugin.author = (pkg.package.publisher) ? pkg.package.publisher.username : null;
         plugin.verifiedPlugin = this.verifiedPlugins.includes(pkg.package.name);
-
+        plugin.icon = this.verifiedPluginsIcons[pkg.package.name]
+          ? `${this.verifiedPluginsIconsPrefix}${this.verifiedPluginsIcons[pkg.package.name]}`
+          : null;
         return plugin;
       });
 
@@ -350,6 +357,7 @@ export class PluginsService {
         description: (pkg.description) ?
           pkg.description.replace(/(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pkg.name,
         verifiedPlugin: this.verifiedPlugins.includes(pkg.name),
+        icon: this.verifiedPluginsIcons[pkg.name],
       } as HomebridgePlugin;
 
       // it's not installed; finish building the response
@@ -364,6 +372,9 @@ export class PluginsService {
       };
       plugin.author = (pkg.maintainers.length) ? pkg.maintainers[0].name : null;
       plugin.verifiedPlugin = this.verifiedPlugins.includes(pkg.name);
+      plugin.icon = this.verifiedPluginsIcons[pkg.name]
+        ? `${this.verifiedPluginsIconsPrefix}${this.verifiedPluginsIcons[pkg.name]}`
+        : null;
 
       return [plugin];
     } catch (e) {
@@ -821,6 +832,9 @@ export class PluginsService {
         manufacturer: {
           type: 'string',
         },
+        firmwareRevision: {
+          type: 'string',
+        },
         model: {
           type: 'string',
         },
@@ -1202,6 +1216,9 @@ export class PluginsService {
       description: (pjson.description) ?
         pjson.description.replace(/(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pjson.name,
       verifiedPlugin: this.verifiedPlugins.includes(pjson.name),
+      icon: this.verifiedPluginsIcons[pjson.name]
+        ? `${this.verifiedPluginsIconsPrefix}${this.verifiedPluginsIcons[pjson.name]}`
+        : null,
       installedVersion: installPath ? (pjson.version || '0.0.1') : null,
       globalInstall: (installPath !== this.configService.customPluginPath),
       settingsSchema: await pathExists(resolve(installPath, pjson.name, 'config.schema.json')) || this.miscSchemas[pjson.name],
@@ -1475,7 +1492,13 @@ export class PluginsService {
     clearTimeout(this.verifiedPluginsRetryTimeout);
     try {
       this.verifiedPlugins = (
-        await this.httpService.get('https://raw.githubusercontent.com/homebridge/verified/latest/verified-plugins.json', {
+        await this.httpService.get(this.verifiedPluginsJson, {
+          httpsAgent: null,
+        }).toPromise()
+      ).data;
+
+      this.verifiedPluginsIcons = (
+        await this.httpService.get(this.verifiedPluginsIconsJson, {
           httpsAgent: null,
         }).toPromise()
       ).data;