+++ ui/src/app/modules/platform-tools/linux/linux.component.ts
@@ -9,6 +9,5 @@ export class LinuxComponent implements OnInit {
 
   constructor() {}
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 }