+++ ui/src/app/modules/plugins/plugins.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
-import { DonateModalComponent } from './donate-modal/donate-modal.component';
 import { PluginCardComponent } from './plugin-card/plugin-card.component';
+import { PluginInfoComponent } from './plugin-card/plugin-info/plugin-info.component';
 import { PluginsRoutingModule } from './plugins-routing.module';
 import { PluginsComponent } from './plugins.component';
 import { CoreModule } from '@/app/core/core.module';
@@ -13,8 +13,8 @@ import { ManagePluginsModule } from '@/app/core/manage-plugins/manage-plugins.mo
 @NgModule({
   declarations: [
     PluginsComponent,
-    DonateModalComponent,
     PluginCardComponent,
+    PluginInfoComponent,
   ],
   imports: [
     CommonModule,