+++ ui/src/app/modules/settings/remove-single-cached-accessory-modal/remove-single-cached-accessory-modal.component.ts
@@ -29,7 +29,7 @@ export class RemoveSingleCachedAccessoryModalComponent implements OnInit {
       this.cachedAccessories = await this.$api.get('/server/cached-accessories').toPromise();
     } catch (e) {
       this.toastr.error(
-        'Accessory cache could not be loaded. You may not have any cached accessories.',
+        this.translate.instant('reset.toast_error_message'),
         this.translate.instant('toast.title_error'),
       );
       this.activeModal.close();
@@ -46,7 +46,7 @@ export class RemoveSingleCachedAccessoryModalComponent implements OnInit {
         cacheFile: item.$cacheFile,
       },
     }).subscribe(
-      async data => {
+      async () => {
         await this.loadCachedAccessories();
 
         this.deleting = null;
@@ -57,7 +57,7 @@ export class RemoveSingleCachedAccessoryModalComponent implements OnInit {
 
         this.toastr.success(this.translate.instant('reset.toast_cached_accessory_removed'), this.translate.instant('toast.title_success'));
       },
-      err => {
+      () => {
         this.deleting = null;
         this.toastr.error(this.translate.instant(
           'reset.toast_failed_to_delete_cached_accessory'),