+++ ui/src/app/modules/status/widgets/cpu-widget/cpu-widget.component.ts
@@ -6,8 +6,8 @@ import {
   OnInit,
   ViewChild,
 } from '@angular/core';
-import { ChartOptions } from 'chart.js';
-import { BaseChartDirective, Color } from 'ng2-charts';
+import { ChartConfiguration } from 'chart.js';
+import { BaseChartDirective } from 'ng2-charts';
 import { Subscription, interval } from 'rxjs';
 import { SettingsService } from '@/app/core/settings.service';
 import { WsService } from '@/app/core/ws.service';
@@ -27,46 +27,46 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
   public cpuTemperature = {} as any;
   public currentLoad = 0;
 
-  public lineChartData = [{ data: [] }];
-  public lineChartLabels = [];
+  public lineChartType: ChartConfiguration['type'] = 'line';
 
-  public lineChartOptions: (ChartOptions & { annotation: any }) = {
-    responsive: true,
-    legend: {
-      display: false,
-    },
-    tooltips: {
-      enabled: false,
-    },
-    scales: {
-      xAxes: [
-        {
-          display: false,
-        },
-      ],
-      yAxes: [
-        {
-          display: false,
-          ticks: {
-            max: 100,
-            min: 0,
-          },
-        },
-      ],
-    },
-    annotation: {
-      annotations: [],
-    },
+  public lineChartData: ChartConfiguration['data'] = {
+    datasets: [{ data: [] }],
   };
 
-  public lineChartColors: Color[] = [
-    { // grey
-      backgroundColor: 'rgba(148,159,177,0.2)',
-      borderColor: 'rgba(148,159,177,0.2)',
-      pointRadius: 0,
-      borderWidth: 1,
+  public lineChartLabels = [];
+
+  public lineChartOptions: ChartConfiguration['options'] = {
+    responsive: true,
+    elements: {
+      point: {
+        radius: 0,
+      },
+      line: {
+        tension: 0.4,
+        backgroundColor: 'rgba(148,159,177,0.2)',
+        borderColor: 'rgba(148,159,177,0.2)',
+        fill: 'origin',
+      },
     },
-  ];
+    plugins: {
+      legend: {
+        display: false,
+      },
+      tooltip: {
+        enabled: false,
+      },
+    },
+    scales: {
+      x: {
+        display: false,
+      },
+      y: {
+        display: false,
+        max: 100,
+        min: 0,
+      },
+    },
+  };
 
   private io = this.$ws.getExistingNamespace('status');
   private intervalSubscription: Subscription;
@@ -77,6 +77,13 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
   ) {}
 
   ngOnInit() {
+    // lookup the chart color based on the current theme
+    const userColor = getComputedStyle(this.widgetBackground.nativeElement).backgroundColor;
+    if (userColor) {
+      this.lineChartOptions.elements.line.backgroundColor = userColor;
+      this.lineChartOptions.elements.line.borderColor = userColor;
+    }
+
     this.io.connected.subscribe(async () => {
       this.getServerCpuInfo();
     });
@@ -90,13 +97,6 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
         this.getServerCpuInfo();
       }
     });
-
-    // lookup the chart color based on the current theme
-    const chartColor = getComputedStyle(this.widgetBackground.nativeElement).backgroundColor;
-    if (chartColor) {
-      this.lineChartColors[0].backgroundColor = chartColor;
-      this.lineChartColors[0].borderColor = chartColor;
-    }
   }
 
   getServerCpuInfo() {
@@ -104,15 +104,17 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
       this.cpuTemperature = data.cpuTemperature;
       this.currentLoad = data.currentLoad;
 
-      if (!this.lineChartData[0].data.length) {
-        this.lineChartData[0].data = data.cpuLoadHistory;
-        this.lineChartLabels = data.cpuLoadHistory.map(x => 'point');
+      if (!this.lineChartData.datasets[0].data.length) {
+        this.lineChartData.datasets[0].data = {
+          ...data.cpuLoadHistory,
+        };
+        this.lineChartLabels = data.cpuLoadHistory.map(() => 'point');
       } else {
-        this.lineChartData[0].data.push(data.currentLoad);
+        this.lineChartData.datasets[0].data.push(data.currentLoad);
         this.lineChartLabels.push('point');
 
-        if (this.lineChartData[0].data.length > 60) {
-          this.lineChartData[0].data.shift();
+        if (this.lineChartData.datasets[0].data.length > 60) {
+          this.lineChartData.datasets[0].data.shift();
           this.lineChartLabels.shift();
           this.chart.update();
         }