+++ ui/src/app/modules/support/support-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { SupportComponent } from './support.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: SupportComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class SupportRoutingModule {}